0.5 Update!
Added in 'Rat', 'Item' and 'Food' classes!  Sprites for the rat and food
have been added as well.

The spirit 'essence' numerical value has been changed into an array to
allow dynamic acquisition.  It now pulls values from a list and adjusts
user input depending on the order the player has acquired the forms in!

Additionally, basic structure for object interactions have been
streamlined, and I have added in a sprite layering function (though it's
unnoticeable because of how sprites do not collide...).