.6 Update
I added in the spirit form abilities and their interactions!  Sprites
were also added and updated to better mesh with the grid and each other.

As a spirit, you can use illuminate to comfort the stag to receive its
blessing.  Additionally, you can pass through animals and objects that
you normally wouldn't be able to when you're in critter form!